Moved UniqueFactorizationDomain derivation to a place where it can be found during implicit search.
